Cleanup architecture by strictly reflecting file structure in parsed abstract file tree, fix some math errors, start changing types to conform to types used in specification and standard implementation - djvulibre

+++ DjvuNet/DjvuNet/Graphics/Pixel.cs
@@ -256,7 +256,7 @@ namespace DjvuNet.Graphics
         /// </returns>
         public override int GetHashCode()
         {
-            return unchecked((int)0xff000000) | (Red << 16) | (Green << 8) | Blue;
+            return unchecked((int)0xff000000) | ((int)Red << 16) | ((int)Green << 8) | (int)Blue;
         }
 
         /// <summary>
